build

+++ scripts/build.sh
@@ -1,19 +1,49 @@
+build_clean=true
+use_polonius=true
+
 # exit when any command fails
 set -e
 
 # to request sudo in the beginning
 sudo uname -r
 
+rust_version="stable"
+if $use_polonius
+then
+  rust_version="nightly"
+fi
+
+rustup install $rust_version
+rustup default $rust_version
+rustup update
+
 cd ..
+
+if $build_clean
+then
 rm -f ./Cargo.lock
 rm -rf ./target
+cargo clean
+fi
 
-rustup update
-rustup default stable
+cd ./src
+CARGO_INCREMENTAL=0
+RUSTFLAGS="-Ctarget-cpu=native"
 
+if $use_polonius
+then
+  RUSTFLAGS="-Z polonius $RUSTFLAGS"
+fi
+
+# release build
 cargo build --release
 
-cd ./scripts
+# polonius debug build
+#cargo +nightly rustc -- -Z polonius
+
+# polonius release build
+#cargo +nightly rustc --release -- -Z polonius
+
+cd ../scripts
 chmod +x ./run.sh
 ./run.sh
-
